src/test/java/enhanced_inventory/server/service/UserServiceTest.java
package enhanced_inventory.server.service;

import enhanced_inventory.server.domain.User;
import enhanced_inventory.server.dto.UserDto;
import enhanced_inventory.server.repository.UserRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.Optional;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.when;

class UserServiceTest {
  @InjectMocks
  private UserService userService;

  @Mock
  private UserRepository userRepository;

  @BeforeEach
  public void init() {
    MockitoAnnotations.openMocks(this);
  }

  @Test
  public void testUpdateUser() {
    User user = new User();
    int userId = 1;
    // ... set other fields

    UserDto userDto = new UserDto(
        "1", "password", "email", "role", "memo", "companyId",
        null, null, null, null
    );

    when(userRepository.findById("1")).thenReturn(Optional.of(user));

    UserDto result = userService.updateUser("1", userDto);

    // assertions
    assertEquals("1", result.userId());
    // ... add more assertions
  }

  @Test
  public void testSaveUser() {
    // Similar to the above, set up your mock and assertions
  }

  @Test
  public void testDeleteUser() {
    // Similar to the above, set up your mock and assertions
  }

  @Test
  public void testGetAllUsers() {
    // Similar to the above, set up your mock and assertions
  }

}